Split out tests of diagnostic functions
